requirements.txt
datasets==3.1.0
flwr==1.9.0
flwr-datasets==0.5.0
hydra-core==1.3.2
numpy==1.26.4
wandb==0.18.7
python-dotenv==1.0.1
torch==2.4.1
slbd @ git+https://github.com/sands-lab/splitbud.git@87d151eb30644aea4035a7407c3695f2000f5465
torchvision==0.19.1
